+++ internal/c/libqb/build.mk
@@ -1,10 +1,16 @@
 
 libqb-objs-y += $(PATH_LIBQB)/src/threading.o
 libqb-objs-y += $(PATH_LIBQB)/src/buffer.o
+libqb-objs-y += $(PATH_LIBQB)/src/command.o
 libqb-objs-y += $(PATH_LIBQB)/src/filepath.o
 libqb-objs-y += $(PATH_LIBQB)/src/filesystem.o
 libqb-objs-y += $(PATH_LIBQB)/src/datetime.o
+libqb-objs-y += $(PATH_LIBQB)/src/error_handle.o
 libqb-objs-y += $(PATH_LIBQB)/src/rounding.o
+libqb-objs-y += $(PATH_LIBQB)/src/qbs.o
+libqb-objs-y += $(PATH_LIBQB)/src/qbs_str.o
+libqb-objs-y += $(PATH_LIBQB)/src/qbs_cmem.o
+libqb-objs-y += $(PATH_LIBQB)/src/string_functions.o
 
 libqb-objs-$(DEP_HTTP) += $(PATH_LIBQB)/src/http.o
 libqb-objs-y$(DEP_HTTP) += $(PATH_LIBQB)/src/http-stub.o
